% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smurf.R
\docType{package}
\name{ssmurf-package}
\alias{ssmurf}
\alias{ssmurf-package}
\title{ssmurf: Sparse Multi-Type Regularized Feature Modeling (package for testing purposes)}
\description{
Adjusted version of the original smurf package. Hopefully the extra s in ssmurf stands for speedy and not slow. The original package is an implementation of the SMuRF algorithm of Devriendt et al. (2021) <doi:10.1016/j.insmatheco.2020.11.010> to fit generalized linear models (GLMs) with multiple types of predictors via regularized maximum likelihood.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/TReynkens/smurf}
  \item Report bugs at \url{https://gitlab.com/TReynkens/smurf/-/issues}
}

}
\author{
\strong{Maintainer}: Tom Reynkens \email{tomreynkens@hotmail.com} (\href{https://orcid.org/0000-0002-5516-5107}{ORCID})

Authors:
\itemize{
  \item Bavo DC Campo \email{bavo.decock@kuleuven.be}
  \item Sander Devriendt \email{sander.devriendt@kuleuven.be}
  \item Katrien Antonio
}

}
